+++ CHANGELOG.md
@@ -9,3 +9,4 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Nothing class with static instance
+- `Task` to `Task<Nothing>` extension method